Updated after second review
Updated Autoware repository setup to create the repository in an explicit directory (~/workspace), then updated all other steps in  the README and tutorial to reference this explicit directory.

Also added a note to step 2 to state that ROS installation may fail and to provide a link to the appropriate manual steps of the ROS Melodic installation guide.