Update TF.md
Cleaned up and rewrote quite a large part of this design doc, and so I'm submitting a PR in a different branch.

I'm not very confident about TF, so it is possible that I may have changed the intended meaning of some sentences, so a careful review would be appreciated.