Readme docs update (#189)
* Initial commit

Proposed changes include review comments from a previously submitted PR.

* Second commit

Made some additional updates to the instructions and included build times, since the pre-requisite stuff takes 45 minutes!

* Update SimulationTutorial.md

* Updated after second review

Updated Autoware repository setup to create the repository in an explicit directory (~/workspace), then updated all other steps in  the README and tutorial to reference this explicit directory.

Also added a note to step 2 to state that ROS installation may fail and to provide a link to the appropriate manual steps of the ROS Melodic installation guide.

* Update README.md

Reworded the description of the repository