Add tutorials from ARM DevSummit hands-on workshop
These tutorials were created for the Autoware workshop that was given during the ARM DevSummit in October 2020.

Since the original hands-on exercises used Autoware Online, I've modified the exercises slightly but they should otherwise be fine.

Assuming these tutorials are approved to be added to the repository, I have a zipped rosbag that needs to be uploaded to the same publicly accessible Google Drive folder used for the other sample maps and rosbag files.